rename UriParser_Solution to UriParser
